coc 서브 라우트 생성, /coc 접근시 첫번째 목차로 리다이렉트하도록 변경
